extracted all high-level functionality out of the ExifTool class.  the ExifToolHelper class currently contains the previous functionality in a verbatim way.  No functionality was changed.
