+++ exiftool/__init__.py
@@ -3,6 +3,7 @@
 # import as directory
 
 # make all of the original exiftool stuff available in this namespace
-from .exiftool import *
+from .exiftool import ExifTool
+from .helper import ExifToolHelper
 
 from .constants import DEFAULT_EXECUTABLE
